check if user has access

+++ src/main/java/com/car_dealership/CarDealership/controllers/AdvertisementController.java
@@ -5,9 +5,12 @@ import com.car_dealership.CarDealership.models.Advertisement;
 import com.car_dealership.CarDealership.models.Car;
 import com.car_dealership.CarDealership.services.AdvertisementService;
 import com.car_dealership.CarDealership.services.CarService;
+import com.car_dealership.CarDealership.services.UserService;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -17,6 +20,7 @@ import java.util.stream.IntStream;
 public class AdvertisementController {
     private AdvertisementService advertisementService;
     private CarService carService;
+    private UserService userService;
 
     @GetMapping("/advertisements")
     public List<AdvertisementDto> getAllAdvertisements() {
@@ -41,15 +45,24 @@ public class AdvertisementController {
     }
 
     @PutMapping("/advertisements/{id}")
-    public AdvertisementDto updateAdvertisement(@PathVariable("id") Advertisement advertisement, @RequestBody AdvertisementDto updatedAdvertisementDto) {
-        Advertisement updatedAdvertisement = advertisementService.updateAdvertisement(advertisement, updatedAdvertisementDto);
-        Car car = updatedAdvertisement.getCar();
-        return AdvertisementDto.createDtoFromAdvertisement(updatedAdvertisement, car);
+    public AdvertisementDto updateAdvertisement(@PathVariable("id") Advertisement advertisement, @RequestBody AdvertisementDto updatedAdvertisementDto)
+            throws AccessDeniedException {
+        if (userService.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName()).getId() == advertisement.getUser().getId()) {
+            Advertisement updatedAdvertisement = advertisementService.updateAdvertisement(advertisement, updatedAdvertisementDto);
+            Car car = updatedAdvertisement.getCar();
+            return AdvertisementDto.createDtoFromAdvertisement(updatedAdvertisement, car);
+        }
+        throw new AccessDeniedException("access denied");
     }
 
     @DeleteMapping("/advertisements/{id}")
-    public void deleteAdvertisement(@PathVariable("id") Advertisement advertisement) {
-        advertisementService.deleteAdvertisement(advertisement);
-        carService.deleteCar(advertisement.getCar());
+    public void deleteAdvertisement(@PathVariable("id") Advertisement advertisement)
+            throws AccessDeniedException {
+        if (userService.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName()).getId() == advertisement.getUser().getId()) {
+            advertisementService.deleteAdvertisement(advertisement);
+            carService.deleteCar(advertisement.getCar());
+        } else {
+            throw new AccessDeniedException("access denied");
+        }
     }
 }
